Access optical counts without memspace
